Moved shared files, dependencies and configuration to `common` module
