+++ server/categories-service/src/main/resources/application.yml
@@ -1,45 +1,3 @@
-server:
-  port: 8020
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health
-  endpoint:
-    health:
-      show-details: always
-  health:
-    db:
-      enabled: true
-    kafka:
-      enabled: true
-
-security:
-  whitelist:
-    - /actuator/**
-
 spring:
-  application:
-    name: categories-service
-  datasource:
-    url: jdbc:postgresql://postgres-categories-service:5432/categories_db
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    defer-datasource-initialization: true
-    hibernate:
-      ddl-auto: create-drop
-  sql:
-    init:
-      mode: always
-  kafka:
-    bootstrap-servers: kafka:9092
-    consumer:
-      group-id: categories-group
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
-      properties:
-        spring.json.trusted.packages: "*"
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+  config:
+    import: classpath:application-common.yml
