+++ server/common/src/main/kotlin/xelagurd/socialdating/server/security/AdminAccess.kt
