--- server/defining-themes-service/src/main/kotlin/xelagurd/socialdating/server/security/HeaderAuthFilter.kt
@@ -1,31 +1,0 @@
-package xelagurd.socialdating.server.security
-
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-
-@Component
-class HeaderAuthFilter : OncePerRequestFilter() {
-
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain
-    ) {
-        val username = request.getHeader("X-Auth-Username")
-        val role = request.getHeader("X-Auth-Role")
-
-        if (username != null && role != null) {
-            val authority = SimpleGrantedAuthority("ROLE_$role")
-            val authentication = UsernamePasswordAuthenticationToken(username, null, listOf(authority))
-            SecurityContextHolder.getContext().authentication = authentication
-        }
-
-        filterChain.doFilter(request, response)
-    }
-}
