+++ server/defining-themes-service/src/main/resources/application.yml
@@ -1,42 +1,3 @@
-server:
-  port: 8030
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health
-  endpoint:
-    health:
-      show-details: always
-  health:
-    db:
-      enabled: true
-    kafka:
-      enabled: true
-
-security:
-  whitelist:
-    - /actuator/**
-
 spring:
-  application:
-    name: defining-themes-service
-  datasource:
-    url: jdbc:postgresql://postgres-defining-themes-service:5432/defining_themes_db
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    defer-datasource-initialization: true
-    hibernate:
-      ddl-auto: create-drop
-  sql:
-    init:
-      mode: always
-  kafka:
-    bootstrap-servers: kafka:9092
-    consumer:
-      group-id: defining-themes-group
-      key-deserializer: org.apache.kafka.common.serialization.StringDeserializer
-      value-deserializer: org.springframework.kafka.support.serializer.JsonDeserializer
-      properties:
-        spring.json.trusted.packages: "*"
+  config:
+    import: classpath:application-common.yml
