--- server/statements-service/src/main/kotlin/xelagurd/socialdating/server/security/AdminAccess.kt
@@ -1,8 +1,0 @@
-package xelagurd.socialdating.server.security
-
-import org.springframework.security.access.prepost.PreAuthorize
-
-@Target(AnnotationTarget.FUNCTION, AnnotationTarget.CLASS)
-@Retention(AnnotationRetention.RUNTIME)
-@PreAuthorize("hasRole('ADMIN')")
-annotation class AdminAccess
