--- server/statements-service/src/main/kotlin/xelagurd/socialdating/server/security/SecurityConfig.kt
@@ -1,37 +1,0 @@
-package xelagurd.socialdating.server.security
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.context.annotation.Profile
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
-import org.springframework.security.web.SecurityFilterChain
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
-
-@Profile("!test")
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity
-class SecurityConfig(
-    private val headerAuthFilter: HeaderAuthFilter,
-    private val securityProperties: SecurityProperties
-) {
-
-    @Bean
-    fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
-        val whitelist = securityProperties.whitelist
-
-        http
-            .csrf { it.disable() }
-            .formLogin { it.disable() }
-            .httpBasic { it.disable() }
-            .addFilterBefore(headerAuthFilter, UsernamePasswordAuthenticationFilter::class.java)
-            .authorizeHttpRequests {
-                it.requestMatchers(*whitelist.toTypedArray()).permitAll()
-                it.anyRequest().authenticated()
-            }
-
-        return http.build()
-    }
-}
