+++ server/statements-service/src/main/resources/application-dev.yml
@@ -1,0 +1,18 @@
+server:
+  port: 8040
+
+management:
+  health:
+    kafka:
+      enabled: true
+
+spring:
+  application:
+    name: statements-service
+  datasource:
+    url: jdbc:postgresql://postgres-statements-service:5432/statements_db
+  kafka:
+    bootstrap-servers: kafka:9092
+    producer:
+      key-serializer: org.apache.kafka.common.serialization.StringSerializer
+      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
