+++ server/statements-service/src/main/resources/application.yml
@@ -1,39 +1,3 @@
-server:
-  port: 8040
-
-management:
-  endpoints:
-    web:
-      exposure:
-        include: health
-  endpoint:
-    health:
-      show-details: always
-  health:
-    db:
-      enabled: true
-    kafka:
-      enabled: true
-
-security:
-  whitelist:
-    - /actuator/**
-
 spring:
-  application:
-    name: statements-service
-  datasource:
-    url: jdbc:postgresql://postgres-statements-service:5432/statements_db
-    driver-class-name: org.postgresql.Driver
-  jpa:
-    defer-datasource-initialization: true
-    hibernate:
-      ddl-auto: create-drop
-  sql:
-    init:
-      mode: always
-  kafka:
-    bootstrap-servers: kafka:9092
-    producer:
-      key-serializer: org.apache.kafka.common.serialization.StringSerializer
-      value-serializer: org.springframework.kafka.support.serializer.JsonSerializer
+  config:
+    import: classpath:application-common.yml
