+++ server/statements-service/src/test/kotlin/xelagurd/socialdating/server/test/StatementsMicroserviceTest.kt
@@ -18,7 +18,7 @@ import xelagurd.socialdating.server.model.details.UserStatementDetails
 import xelagurd.socialdating.server.model.enums.StatementReactionType.FULL_MAINTAIN
 import xelagurd.socialdating.server.utils.TestUtils.toRequestParams
 
-@ActiveProfiles("test")
+@ActiveProfiles("dev", "test")
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @Import(NoSecurityConfig::class)
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
