+++ server/users-service/src/main/kotlin/xelagurd/socialdating/server/exception/AuthExceptionHandler.kt
@@ -1,0 +1,33 @@
+package xelagurd.socialdating.server.exception
+
+import org.springframework.http.HttpStatus
+import org.springframework.security.core.AuthenticationException
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.ResponseStatus
+import org.springframework.web.bind.annotation.RestControllerAdvice
+import io.github.oshai.kotlinlogging.KotlinLogging
+import io.jsonwebtoken.JwtException
+import xelagurd.socialdating.server.utils.ExceptionUtils.getErrorPositionFromStackTrace
+
+@RestControllerAdvice
+class AuthExceptionHandler {
+    val logger = KotlinLogging.logger { }
+
+    @ExceptionHandler(AuthenticationException::class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    fun handleAuthenticationException(ex: AuthenticationException): String {
+        val message = ex.message ?: "Unauthorized"
+        val origin = getErrorPositionFromStackTrace(ex.stackTrace)
+        logger.error { "Class: ${ex.javaClass.simpleName}, origin: $origin, message: $message" }
+        return message
+    }
+
+    @ExceptionHandler(JwtException::class)
+    @ResponseStatus(HttpStatus.UNAUTHORIZED)
+    fun handleJwtException(ex: JwtException): String {
+        val message = ex.message ?: "Invalid JWT token"
+        val origin = getErrorPositionFromStackTrace(ex.stackTrace)
+        logger.error { "Class: ${ex.javaClass.simpleName}, origin: $origin, message: $message" }
+        return message
+    }
+}
