+++ server/users-service/src/main/kotlin/xelagurd/socialdating/server/security/AuthSecurityConfig.kt
@@ -5,25 +5,16 @@ import org.springframework.context.annotation.Configuration
 import org.springframework.context.annotation.Profile
 import org.springframework.security.authentication.AuthenticationManager
 import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity
-import org.springframework.security.config.annotation.web.builders.HttpSecurity
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity
 import org.springframework.security.core.userdetails.UserDetailsService
 import org.springframework.security.core.userdetails.UsernameNotFoundException
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
 import org.springframework.security.crypto.password.PasswordEncoder
-import org.springframework.security.web.SecurityFilterChain
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter
 import xelagurd.socialdating.server.repository.UsersRepository
 
 @Profile("!test")
 @Configuration
-@EnableWebSecurity
-@EnableMethodSecurity
-class SecurityConfig(
-    private val usersRepository: UsersRepository,
-    private val headerAuthFilter: HeaderAuthFilter,
-    private val securityProperties: SecurityProperties
+class AuthSecurityConfig(
+    private val usersRepository: UsersRepository
 ) {
 
     @Bean
@@ -40,22 +31,5 @@ class SecurityConfig(
         authenticationConfiguration: AuthenticationConfiguration
     ): AuthenticationManager {
         return authenticationConfiguration.authenticationManager
-    }
-
-    @Bean
-    fun securityFilterChain(http: HttpSecurity): SecurityFilterChain {
-        val whitelist = securityProperties.whitelist
-
-        http
-            .csrf { it.disable() }
-            .formLogin { it.disable() }
-            .httpBasic { it.disable() }
-            .addFilterBefore(headerAuthFilter, UsernamePasswordAuthenticationFilter::class.java)
-            .authorizeHttpRequests {
-                it.requestMatchers(*whitelist.toTypedArray()).permitAll()
-                it.anyRequest().authenticated()
-            }
-
-        return http.build()
     }
 }
