--- server/users-service/src/main/kotlin/xelagurd/socialdating/server/security/SecurityProperties.kt
@@ -1,10 +1,0 @@
-package xelagurd.socialdating.server.security
-
-import org.springframework.boot.context.properties.ConfigurationProperties
-import org.springframework.stereotype.Component
-
-@Component
-@ConfigurationProperties(prefix = "security")
-data class SecurityProperties(
-    var whitelist: List<String> = emptyList()
-)
